+++ src/handlers/tsonUnknown.test.ts
@@ -1,0 +1,47 @@
+import { assert, expect, test } from "vitest";
+import { expectError } from "../testUtils.js";
+import { createTupleson } from "../tson.js";
+import { tsonSet } from "./tsonSet.js";
+import { UnknownObjectGuardError, tsonUnknown } from "./tsonUnknown.js";
+
+test("guard unwanted objects", () => {
+	// Sets are okay, but not Maps
+	const t = createTupleson({
+		types: {
+			Set: tsonSet,
+			// defined last so it runs last
+			unknownObjectGuard: tsonUnknown,
+		},
+	});
+
+	{
+		// sets are okay
+		const expected = new Set([1]);
+
+		const stringified = t.stringify(expected);
+		const deserialized = t.parse(stringified);
+
+		expect(deserialized).toEqual(expected);
+	}
+
+	{
+		// plain objects are okay
+		const expected = { a: 1 };
+		const stringified = t.stringify(expected);
+		const deserialized = t.parse(stringified);
+		expect(deserialized).toEqual(expected);
+	}
+
+	{
+		// maps are not okay
+		const expected = new Map([["a", 1]]);
+
+		const err = expectError(() => t.parse(t.stringify(expected)));
+		assert(err instanceof UnknownObjectGuardError);
+
+		expect(err).toMatchInlineSnapshot(
+			"[UnknownObjectGuardError: Unknown object found]",
+		);
+		expect(err.value).toEqual(expected);
+	}
+});
